requirements.txt
boto3>=1.28.0
kafka-python>=2.0.2
